+++ src/batch-network-interface.js
@@ -4,7 +4,7 @@ import { extractRequestFiles } from './helpers'
 export class HTTPUploadBatchNetworkInterface extends HTTPBatchedNetworkInterface {
   batchedFetchFromRemoteEndpoint({ requests, options }) {
     // Skip upload proccess if SSR
-    if (typeof window !== 'undefined') {
+    if (typeof FormData !== 'undefined') {
       // Extract any files from the request
       const batchFiles = []
       const batchOperations = requests.map((request, operationIndex) => {
@@ -26,7 +26,7 @@ export class HTTPUploadBatchNetworkInterface extends HTTPBatchedNetworkInterface
         })
 
         // Build the form
-        const formData = new window.FormData()
+        const formData = new FormData()
         formData.append('operations', JSON.stringify(batchOperations))
         batchFiles.forEach(({ operationIndex, files }) => {
           files.forEach(({ variablesPath, file }) =>
@@ -35,7 +35,7 @@ export class HTTPUploadBatchNetworkInterface extends HTTPBatchedNetworkInterface
         })
 
         // Send request
-        return window.fetch(this._uri, {
+        return fetch(this._uri, {
           method: 'POST',
           body: formData,
           ...options